Refactor CMS example to make it easier to understand (#61)
